pyproject.toml
[tool.poetry]
name = "papyrus"
version = "0.1.0"
description = ""
authors = ["ivansantiagojr <ivansantiago.junior@gmail.com>"]
readme = "README.md"
packages = [{ include = "src" }]


[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.0"
uvicorn = "^0.23.1"
sqlalchemy = "^2.0.19"
pydantic-settings = "^2.0.2"
alembic = "^1.11.1"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
python-multipart = "^0.0.6"
psycopg = { extras = ["binary"], version = "^3.1.10" }


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
taskipy = "^1.11.0"
blue = "^0.9.1"
ruff = "^0.0.278"
httpx = "^0.24.1"
isort = "^5.12.0"
factory-boy = "^3.3.0"
freezegun = "^1.2.2"


[tool.ruff]
line-length = 79
exclude = ['.venv', 'migrations']

[tool.isort]
profile = "black"
line_length = 79

[tool.pytest.ini_options]
pythonpath = "."


[tool.blue]
extend-exclude = '(migrations/)'

[tool.taskipy.tasks]
lint = 'ruff . && blue --check . --diff'
format = 'blue .  && isort .'
run = 'uvicorn src.main:app --reload'
# pre_test = 'task lint'
test = 'pytest -s -x --cov=src -vv'
post_test = 'coverage html'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
